+++ src/main/java/blackjack/controller/BlackJackController.java
@@ -3,10 +3,10 @@ package blackjack.controller;
 import java.util.List;
 
 import blackjack.domain.BlackJack;
-import blackjack.domain.ProfitResult;
 import blackjack.domain.gamer.PlayerGroup;
 import blackjack.domain.gamer.role.Player;
 import blackjack.domain.result.GameResult;
+import blackjack.domain.result.ProfitResult;
 import blackjack.dto.DealerResultDto;
 import blackjack.dto.GamerCardsDto;
 import blackjack.dto.GamerCardsResultDto;
