+++ src/test/java/blackjack/domain/BlackJackTest.java
@@ -19,6 +19,7 @@ import blackjack.domain.result.DealerResult;
 import blackjack.domain.result.GameResult;
 import blackjack.domain.result.Match;
 import blackjack.domain.result.PlayerResult;
+import blackjack.domain.result.ProfitResult;
 
 class BlackJackTest {
     @Test
