+++ src/test/java/blackjack/domain/card/CardsTest.java
@@ -1,16 +1,16 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
-import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.*;
 
-import blackjack.domain.card.Card;
-import blackjack.domain.card.CardGroup;
-import blackjack.domain.card.property.CardNumber;
-import blackjack.domain.card.property.CardShape;
 import java.util.stream.Stream;
+
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
+
+import blackjack.domain.card.property.CardNumber;
+import blackjack.domain.card.property.CardShape;
 
 class CardsTest {
