+++ src/test/java/blackjack/domain/gamer/DealerTest.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.gamer;
 
 import static org.assertj.core.api.Assertions.*;
